luci-app-softethervpn: update provided links for SoftEtherVPN clients
I updated the link for downloading Windows client to download the latest
currently available version instead of using beta version, which is not
good to use in production.

Also, the macOS link led to the OpenVPN server, which was wrong. It should
be a client. It was updated to the latest version as well.