bcm27xx-userland: update to latest version
The latest version contains some fixes and additions.

Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>

bcm27xx-userland: update to latest version

Properly recognise all BCM2711 variants

Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>

bcm27xx-userland: factor out a -dev package

Installing headers and static libraries to the target system seems
to be not required for most use cases, so let's factor them
out into a dedicated -dev package.

This cuts down to disk usage to around 50% of the original
package to ~ 2MB - not that disk space is an issue normally,
but when using inside an initramfs only project, it counts.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>

Co-authored-by: Michael Heimpold <mhei@heimpold.de>