ead: fix compilation with GCC14
Trying to compile EAD with now default GCC14 will fail due to now erroring
out on implicit int type as well as implicit function declarations.

Due to this, the packaged configure script will fail on the simple compiler
test as the generated test uses main loop with an implicit int type.

So, instead of patching multiple test cases in the shipped configure script
for tinysrp lets run autoreconf on it so its regenerated.

We also need to pass -Wno-error=implicit-function-declaration as there are
multiple instances of tinysrp code relying on implicit function declarations.

Signed-off-by: Robert Marko <robimarko@gmail.com>