loongarch64: built-in PCI DWMAC
Add the CONFIG_DWMAC_LOONGSON driver to the loongarch64 kernel
configuration as a built-in module.

This driver is required for the integrated Ethernet controller found on
modern Loongson devices such as the 3A6000 NUC and 2K3000. Including it
by default ensures these popular devices have network functionality
out-of-the-box.

Configuration changes were made via `make kernel_menuconfig` for the
loongarch64 target.

Signed-off-by: Vincent Li <vincent.mc.li@gmail.com>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>